leave warning
implemented the functionality of providing a warning if the last user in the room clicks leave button. there will be a pop up which will say that u are the last user and clickin on leave all the content of the room will be deleted permanently so do u want to download the content or not?
the current handle dwnload function which is written will save the file shown on the editor into  a .txt file